# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Telegram File Explorer API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type AuthStatusStep = typeof AuthStatusStep[keyof typeof AuthStatusStep];


export const AuthStatusStep = {
  unauthenticated: 'unauthenticated',
  code_sent: 'code_sent',
  two_fa_required: 'two_fa_required',
  authenticated: 'authenticated',
} as const;

export interface AuthStatus {
  authenticated: boolean;
  step: AuthStatusStep;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  username?: string | null;
  /** @nullable */
  first_name?: string | null;
}

export interface PhoneInput {
  phone: string;
}

export interface CodeInput {
  phone: string;
  code: string;
}

export interface PasswordInput {
  password: string;
}

export interface Channel {
  id: number;
  title: string;
  /** @nullable */
  username?: string | null;
  type: string;
  /** @nullable */
  member_count?: number | null;
  /** @nullable */
  file_count?: number | null;
  /** @nullable */
  total_size?: number | null;
  /** @nullable */
  last_sync?: string | null;
  has_profile_photo: boolean;
}

export interface ChannelList {
  channels: Channel[];
  total: number;
  page: number;
  page_size: number;
}

export interface ChannelStats {
  total_channels: number;
  total_files: number;
  total_size: number;
  /** @nullable */
  last_sync?: string | null;
}

export interface TelegramFile {
  message_id: number;
  channel_id: number;
  filename: string;
  extension: string;
  size: number;
  mime_type: string;
  date: string;
  is_document: boolean;
}

export interface FileList {
  files: TelegramFile[];
  total: number;
  page: number;
  page_size: number;
  total_size: number;
}

export type FileStatsFileTypes = {[key: string]: number};

export interface FileStats {
  total_files: number;
  total_size: number;
  synced: boolean;
  /** @nullable */
  last_sync?: string | null;
  file_types?: FileStatsFileTypes;
}

export interface DownloadRequest {
  channel_id: number;
  /** @nullable */
  message_ids?: number[] | null;
  /** @nullable */
  archive_name?: string | null;
}

export type DownloadJobStatus = typeof DownloadJobStatus[keyof typeof DownloadJobStatus];


export const DownloadJobStatus = {
  pending: 'pending',
  running: 'running',
  completed: 'completed',
  failed: 'failed',
  cancelled: 'cancelled',
} as const;

export interface DownloadJob {
  job_id: string;
  status: DownloadJobStatus;
  channel_id: number;
  total_files: number;
  downloaded_files: number;
  failed_files: number;
  total_size: number;
  downloaded_size: number;
  /** @nullable */
  error?: string | null;
  created_at: string;
  /** @nullable */
  completed_at?: string | null;
  /** @nullable */
  download_url?: string | null;
}

export interface SyncStatus {
  status: string;
  message: string;
  /** @nullable */
  task_id?: string | null;
}

export type ListChannelsParams = {
search?: string;
type?: string;
sort?: string;
order?: string;
page?: number;
page_size?: number;
};

export type ListChannelFilesParams = {
search?: string;
extension?: string;
sort?: string;
order?: string;
page?: number;
page_size?: number;
};